Implement 1-norm vector function
This commit also adds the absolute value function for vectors. Instead
of relying on `stdlib.h` for the implementation of `abs`, it was
implemented in "common_math.h". This avoids a dependency and
(theoretically) reduces the size of the executable.

Additionally, the magnitude function has been renamed to `2norm` to keep
the naming scheme consistent.

NOTE:

- It would be a good idea to find a better name for the `mag_squared`
  function, or add the alias `mag` for `2norm`.